moved search components into search folder
